Fix INT4/i64 mismatch and update E2E tests for event log
SELECT 1 returns INT4; decode as i32 not i64 to fix type mismatch
in author delete path.

Update E2E tests: rename historyId→eventId, restoreBook/Author
parameter history_id→event_id, update assertions for post-state
semantics (update events now record post-update state), and fix
author delete event assertion (name is now null, not pre-delete value).

+++ src/infrastructure/author_repository.rs
@@ -202,7 +202,7 @@ impl AuthorRepository for PgAuthorRepository {
         let mut tx = self.pool.begin().await?;
 
         // Lock the author row to prevent concurrent inserts into book_author after the count check.
-        let exists: Option<(i64,)> =
+        let exists: Option<(i32,)> =
             sqlx::query_as("SELECT 1 FROM author WHERE id = $1 AND user_id = $2 FOR UPDATE")
                 .bind(author_id.to_uuid())
                 .bind(user_id.as_str())
